refactor: move getCurrentDate function to date.utils.ts for better organization
- Removed the getCurrentDate function from utils.ts and relocated it to date.utils.ts in the kronos-core library.
- Updated imports in builder.ts to source getCurrentDate from the new location, ensuring consistent usage across the application.
- Enhanced the getCurrentDate function with detailed documentation, providing examples and clarifying its output format.
