+++ apps/server/src/agents/kronos/utils.ts
@@ -22,15 +22,3 @@ export function extractToolCalls(message: any): any[] {
   }
   return [];
 }
-
-/**
- * Format current date for prompts
- */
-export function getCurrentDate(): string {
-  return new Date().toLocaleDateString('en-US', {
-    weekday: 'long',
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric',
-  });
-}
